Weather Classification Code
The code is functional. Before running the models, I do some cleaning for the original us_accidents_weather_data.csv. That will probably need to change once the data preprocessing is all complete. Each model has the option to be run with over/under sampling to address data imbalance. Additionally, there is an option to increase the amount of data that is used for each model (e.g. 50000 rows vs 5000). Some of the files may require some package installations.